refactoring project structure (actions, constants, scripts, util)
