+++ repo/src/main/java/com/parashift/onlyoffice/scripts/ConverterTest.java
@@ -1,4 +1,4 @@
-package com.parashift.onlyoffice;
+package com.parashift.onlyoffice.scripts;
 
 import java.io.IOException;
