+++ repo/src/main/java/com/parashift/onlyoffice/util/UtilDocConfig.java
@@ -1,6 +1,5 @@
-package com.parashift.onlyoffice;
+package com.parashift.onlyoffice.util;
 
-import org.alfresco.model.ContentModel;
 import org.alfresco.repo.i18n.MessageService;
 import org.alfresco.service.cmr.coci.CheckOutCheckInService;
 import org.alfresco.service.cmr.favourites.FavouritesService;
@@ -11,7 +10,6 @@ import org.alfresco.service.cmr.security.AccessStatus;
 import org.alfresco.service.cmr.security.PermissionService;
 import org.alfresco.service.cmr.security.PersonService;
 import org.alfresco.service.cmr.security.PersonService.PersonInfo;
-import org.json.JSONArray;
 import org.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
